% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ecdf.R
\name{as_ecdf}
\alias{as_ecdf}
\title{Create An ECDF List}
\usage{
as_ecdf(type, ecdf)
}
\arguments{
\item{type}{this is either `shallow` or `deep`}

\item{ecdf}{a data.frame}
}
\value{
A list with three named objects (`type`,`percent_time`,`ecdf`) of
    class `wcECDF`
}
\description{
This creates a standardized list for holding ECDF data and sets the class
to `ecdf`. The S3 class, *wcECDF* consists of three named objects: `type`,
`percent_time`, and `ecdf`. The `type` object is the same as that specified
for the `type` parameter. The `percent_time` object holds the proportion of
the summary period the tag spent within the `shallow` or `deep` portion of
the water column. Lastly, the `ecdf` object is a data.frame with a column,
`ecd_prop`, that specifies the proportional values for the ECDF and a
column, `depth_break`, that reports the corresponding depth value (in meters).
}
