\name{autoplot.HistoPercentDry}
\alias{autoplot.HistoPercentDry}
\title{autoplot.HistoPercentDry}
\usage{
  autoplot.HistoPercentDry(d, ..., xlab = "", ylab = "",
    title = "Percent Dry Timeline")
}
\arguments{
  \item{d}{dataframe object output from MeltHistos
  function}
}
\value{
  ggplot
}
\description{
  autoplot.HistoPercentDry
}
\note{
  future plans to evolve this to an autoplot function
}
\author{
  Josh M London \email{josh.london@noaa.gov}
}

