% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ecdf.R
\name{combine_ecdf}
\alias{combine_ecdf}
\title{Combine Two `wcECDF` Objects Into a Single ECDF}
\usage{
combine_ecdf(ecdf1, ecdf2, return = "pdf")
}
\arguments{
\item{ecdf1}{object of class `wcECDF`; typically the *shallow* region}

\item{ecdf2}{object of class `wcECDF`; typically the *deep* region}
}
\value{
four-column data.frame with columns `ecd_prop`, `pdf`, `prob`, and `depth_break`
}
\description{
Combine Two `wcECDF` Objects Into a Single ECDF
}
