% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdt.R
\docType{data}
\name{.orders}
\alias{.orders}
\title{Read and Tidy PDT Data File}
\format{
An object of class \code{character} of length 4.
}
\usage{
.orders
}
\arguments{
\item{pdt_csv}{file path for the `*-PDTs.csv`}
}
\value{
A tibble
}
\description{
All PDT data are
stored within a `*-PDTs.csv` file that is output from either the
Wildlife Computers Data Portal or DAP processing software. The function
presumes the `*-PDTs.csv` data file is provided as-is from these sources
and has not been edited. The resulting output is a `tibble` that
adheres to tidy data principles and includes five columns (`deploy_id`,
`date`, `depth`, `min_te`, and `max_te`.
}
\keyword{datasets}
