% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format_bins.R
\name{format_bins}
\alias{format_bins}
\title{format bin labels}
\usage{
format_bins(bins)
}
\arguments{
\item{bins}{a vector of bin labels to be formated}
}
\value{
a vector of formated bin labels
}
\description{
format bin labels
}

