% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_names.R
\name{make_names}
\alias{make_names}
\title{Expanded make.names function for creating consistent column names}
\usage{
make_names(x)
}
\arguments{
\item{x}{data frame with columns to be renamed}
}
\value{
a data frame
}
\description{
Expanded make.names function for creating consistent column names
}
