\name{plot_HourlyTimeline}
\alias{plot_HourlyTimeline}
\title{plot_HourlyTimeline}
\usage{
  plot_HourlyTimeline(d)
}
\arguments{
  \item{d}{dataframe object output from MeltHistos
  function}
}
\value{
  ggplot
}
\description{
  Plot hourly timeline data
}
\note{
  future plans to evolve this to an autoplot function
}
\author{
  Josh M London \email{josh.london@noaa.gov}
}

