% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_behav.R
\name{read_behav}
\alias{read_behav}
\title{Parse a *-Behavior.csv file into a proper data.frame}
\usage{
read_behav(behav_file, to_lower = TRUE)
}
\arguments{
\item{behav_file}{file path or file connection to a *-Histos.csv file}

\item{to_lower}{whether to convert the column names to lower case}
}
\value{
a data frame
}
\description{
Parse a *-Behavior.csv file into a proper data.frame
}

