% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastGPS.R
\name{read_fastGPS}
\alias{read_fastGPS}
\title{Parse a *-FastGPS.csv files into a proper data.frame}
\usage{
read_fastGPS(gps_file, to_lower = TRUE, fix_csv = FALSE)
}
\arguments{
\item{gps_file}{file path or file connection to a *-FastGPS.csv file}

\item{to_lower}{whether to convert the column names to lower case}

\item{fix_csv}{whether to attemtp to fix any comma, csv issues}
}
\value{
a data frame
}
\description{
Parse a *-FastGPS.csv files into a proper data.frame
}
