% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastGPS.R
\name{read_gps}
\alias{read_gps}
\title{Parse a *-FastGPS.csv files into a proper data.frame}
\usage{
read_gps(gps_file, to_lower = TRUE)
}
\arguments{
\item{gps_file}{file path or file connection to a *-FastGPS.csv file}

\item{to_lower}{whether to convert the column names to lower case}
}
\value{
a data frame
}
\description{
Parse a *-FastGPS.csv files into a proper data.frame
}

