% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_locs.R
\name{read_locs}
\alias{read_locs}
\title{Parse a *-Locations.csv files into a proper data.frame}
\usage{
read_locs(loc_file, to_lower = TRUE)
}
\arguments{
\item{loc_file}{file path or file connection to a *-Histos.csv file}

\item{to_lower}{whether to convert the column names to lower case}
}
\value{
a data frame
}
\description{
Parse a *-Locations.csv files into a proper data.frame
}

