% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_ecdf.R
\name{spline_ecdf}
\alias{spline_ecdf}
\title{approximate ecdf with spline}
\usage{
spline_ecdf(ecdf, bin.width)
}
\arguments{
\item{ecdf}{}

\item{bin.width}{}
}
\description{
approximate ecdf with spline
}
