% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tidyTimelines.R
\name{tidyTimelines}
\alias{tidyTimelines}
\title{Apply 'tidy' data principles to the timeline histogram data stream}
\usage{
tidyTimelines(histos)
}
\arguments{
\item{histos}{a list returned from \code{read_histos}}
}
\value{
a data frame with tidy, narrow data structure and actual time periods in place of bins
}
\description{
\code{tidyTimelines} returns a 'tidy'd' data frame of timeline data
}
\details{
The histogram data stream is provided in a 'wide' format (each row represents a time period and the observed values are provided in 1 to 72 'bin' columns). This format can be difficult to work with in R and other data analysis platforms (e.g. database tables), so we use the \code{tidyr} and \code{dplyr} packages to manipulate the data into a more flexible, 'narrow' format. This results in a data structure where every row represents a single observation.

This is implemented, here, with the timeline data. For timeline data, tag 'dryness' is provided as either a percentage of each hour the tag was dry or as a binary (1 or 0) value representing whether a tag was dry for a majority of a given 20-minute period. For both of these situations, the values for the 'bin' columns are predictable and we can, in addition to tidying the data structure, also turn the bin values into actual time periods
}

