% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pdf.R
\name{to_pdf}
\alias{to_pdf}
\title{Transform a `wcECDF` Object Into a Probability Density Function}
\usage{
to_pdf(ecdf)
}
\arguments{
\item{ecdf}{two-column data.frame with `ecdf_prop` and `depth_break`}
}
\value{
a two-column data.frame with `ecdf_prop` and `depth_break`
}
\description{
Transform a `wcECDF` Object Into a Probability Density Function
}
