% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcGetDeployID.R
\name{wcGetDeployID}
\alias{wcGetDeployID}
\title{Return a vector of data portal unique IDs associated with a given DeployID}
\usage{
wcGetDeployID(xml_content, deployid = NULL)
}
\arguments{
\item{xml_content}{XML content/data returned from wcPOST (with 'action=get_deployments')}

\item{deployid}{valid deployid character (required)}
}
\value{
a list with ids (a vector of deployment ids) and a df (data frame of deployment summaries)
}
\description{
\code{wcGetDeployID} returns a vector of data portal unique ID(s)
}
\details{
This function presumes a DeployID has been setup for deployments on the Wildife
Computers Data Portal. The vector of deployment ID(s) returned will be a subset that match the
deployid character in the function call. The list returned will also include a simple
data frame with summary information one can use to determine the appropriate id.
}
