% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wcGetIDs.R
\name{wcGetIDs}
\alias{wcGetIDs}
\title{Return a vector of deployment IDs}
\usage{
wcGetIDs(xml_content, xpath = NULL)
}
\arguments{
\item{xml_content}{XML content/data returned from wcPOST (with 'action=get_deployments')}

\item{xpath}{additional customization possible by passing an xpath statement}
}
\value{
returns a vector of deployment IDs
}
\description{
\code{wcGetIDs} returns a vector of deployment IDs
}
\details{
Each deployment in the Wildlife Computers Data Portal is identified by a unique alpha-numeric value. This function searches the XML response data and extracts those IDs
}

