% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wcGetProjectIDs}
\alias{wcGetProjectIDs}
\title{Return a vector of deployment IDs associated with a given Project}
\usage{
wcGetProjectIDs(xml_content, project = NULL)
}
\arguments{
\item{xml_content}{XML content/data returned from wcPOST (with 'action=get_deployments')}

\item{project}{valid project name (required)}
}
\value{
returns a vector of deployment IDs
}
\description{
\code{wcGetProjectIDs} returns a vector of deployment IDs
}
\details{
This function presumes a custom label,'Project', has been setup for deployments on the Wildife
Computers Data Portal. The vector of deployment IDs returned will be a subset that match the
project name provided in the function call.
}

