% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcGetPttID.R
\name{wcGetPttID}
\alias{wcGetPttID}
\title{Return a vector of deployment ID associated with a given PTT}
\usage{
wcGetPttID(xml_content, ptt = NULL)
}
\arguments{
\item{xml_content}{XML content/data returned from wcPOST (with 'action=get_deployments')}

\item{ptt}{valid ptt integer (required)}
}
\value{
a list with ids (a vector of deployment ids) and a df (data frame of deployment summaries)
}
\description{
\code{wcGetPttID} returns a vector of deployment ID(s)
}
\details{
This function presumes a PTT has been setup for deployments on the Wildife
Computers Data Portal. The vector of deployment ID(s) returned will be a subset that match the
ptt integer provided in the function call. The list returned will also include a simple
data frame with summary information one can use to determine the appropriate id.
}
