% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcGetRecentIDs.R
\name{wcGetRecentIDs}
\alias{wcGetRecentIDs}
\title{Return a vector of deployment IDs with new data in the last \code{n} days}
\usage{
wcGetRecentIDs(xml_content, days = 14)
}
\arguments{
\item{xml_content}{XML content/data returned from wcPOST (with 'action=get_deployments')}

\item{days}{integer value specifying the time window from \code{now()} in days}
}
\value{
returns a vector of deployment IDs
}
\description{
\code{wcGetProjectIDs} returns a vector of deployment IDs
}
\details{
This returns a subset of deployment IDs with new data available on the portal within the last
\code{n} days. The default value is for 14 days
}
