% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcGetZip.R
\name{wcGetZip}
\alias{wcGetZip}
\title{Retrieve a single zip file from Wildlife Computers Data Portal}
\usage{
wcGetZip(id, keyfile = NULL)
}
\arguments{
\item{id}{a single character representing a data portal unique deployment identifier}
}
\value{
a path to the zip file
}
\description{
\code{wcGetZip} will return a path to a downloaded zip file
}
\details{
The Wildlife Computers Data Portal will return deployment data in the form of a zipped file with 
various comma-separated files and other accessory files.
}

