% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wcProcessGPS}
\alias{wcProcessGPS}
\title{Process GPS/fastloc location data}
\usage{
wcProcessGPS(xml_content)
}
\arguments{
\item{xml_content}{XML content/data returned from wcPOST (with 'action=get_deployments')}
}
\value{
nothing is returned
}
\description{
\code{wcProcessGPS} determines the GPS/Fastloc tags in need of processing and sends a command to
initiate processing
}

