\docType{package}
\name{wcUtils-package}
\alias{wcUtils}
\alias{wcUtils-package}
\title{wcUtils: Makes life easier for those working with Wildlife Computers data}
\description{
  Makes life easier for those working with Wildlife
  Computers data
}
\details{
  The aim of `wcUtils` is to make your life as an analyst
  of Wildlife Computers data a wee bit easier. You do need
  a rudimentary understanding of R and the package does
  require the R package 'reshape' (available on CRAN). At
  this point, the purpose of \code{wcUtils} is to make your
  life easier, not to do the analysis for you. As such, the
  tools focus more on manipulation of data and less on
  analysis.
}
\author{
  Josh London \email{josh.london@noaa.gov}
}

