% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wcUtils.R
\docType{package}
\name{wcUtils}
\alias{wcUtils}
\alias{wcUtils-package}
\title{wcUtils: A package for import/export of Wildlife Computers tag data}
\description{
wcUtils provides functionality for working with data from select
(SPLASH, SPOT) Wildlife Computers satellite telemetry tags. The package
relies on data files produced by the Wildlife Computers DAP program or
downloaded from the Wildlife Computers Data Portal. Data are read into R
}
\section{wcUtils functions}{

  \itemize{
    \item read_behav
    \item read_histos
    \item read_locs
    \item tidyDiveDepths
    \item tidyDiveDurations
    \item tidyTimeAtDepth
    \item tidyTimelines
    \item wcGetDownload
    \item wcGetIDs
    \item wcGetProjectIDs
    \item wcGetRecentIDs
    \item wcPOST
  }
}

