% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_pct_depth.R
\name{x_pct_depth}
\alias{x_pct_depth}
\title{Determine percent TAD at specific depth}
\usage{
x_pct_depth(ecdf, pct_tad)
}
\arguments{
\item{ecdf}{}

\item{pct_tad}{}
}
\description{
Determine percent TAD at specific depth
}
