% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_behav.R
\name{read_behav}
\alias{read_behav}
\title{Parse a *-Behavior.csv file into a proper data.frame}
\usage{
read_behav(behav_file, to_lower = TRUE, fix_csv = FALSE)
}
\arguments{
\item{behav_file}{file path or file connection to a *-Behavior.csv file}

\item{to_lower}{whether to convert the column names to lower case}

\item{fix_csv}{whether to attemtp to fix any comma, csv issues}
}
\value{
a data frame
}
\description{
Parse a *-Behavior.csv file into a proper data.frame
}
