% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcGetZip.R
\name{wcGetZip}
\alias{wcGetZip}
\title{Retrieve a single zip file from Wildlife Computers Data Portal}
\usage{
wcGetZip(
  id,
  wc.key = Sys.getenv("wcAccessKey"),
  wc.secret = Sys.getenv("wcSecretKey"),
  keyfile = NULL
)
}
\arguments{
\item{wc.key}{public access key (default retrieves from option value set in .Renviron)}

\item{wc.secret}{secret access key (default retrieves from option value set in .Renviron)}

\item{keyfile}{path to a json formatted keyfile with wcAccessKey and wcSecretKey}
}
\value{
a path to the zip file
}
\description{
\code{wcGetZip} will return a path to a downloaded zip file
}
\details{
The Wildlife Computers Data Portal will return deployment data in the form of a zipped file with 
various comma-separated files and other accessory files.
}
